Fixed environment export setting for virtualenv
This is a band-aid for a fix for ensuring environment variables
are exported properly when directly activating the virtualenv. The issue
is that the gunicorn job doesnt actually activate the virtualenv but
is calling the gunicorn binary from within the env folder. I had to
utilize the EnvironmentFile parameter in systemd which did not take in
variable lines that had `export` in front of them.