Break django manage tasks into three stages
This applies to django manage tasks, I expanded the default behavior
on the database set so that `collectstatic` and `migrate` should be run
everytime. Also allows optional pre db tasks and post db tasks to be added.
These are designed to only be run once per install (detected by when the
virtualenv folder is created).