Allow over-ride of pip and npm package detection
Typically we were relying on Git pull changes on the django code
to determine when we instal NPM + Virtualenv packages. Ran into a recent
situation where we disabled the git pull task so the node packages
werent getting properly installed. This provides a new variable
(defaults to false) that will force attempt to install npm and virtualenv
packages.