Allow override of django settings file
This allows a playbook to specify a local settings file that would ideally
over-ride settings files in the repo. Typically this plays nice when the repo
you are deploying has an ignore exception for a settings file. This is usually
best practice on a django repo to ignore a local.py settings file for developers
to hack on.