Refactor role to build locally
This is a pretty substantial commit - the result of a big rebase during a
long and boring trial-test-repeat phase. Anywhoo - this commit dramatically
changes the logic of the role. Where as before the "compilation" of the
virtualenv and node logic was being done on the server, now that logic is being
completed client-side under a docker container. Because of this, this commit
severely breaks backwards compatibility with playbooks using this role before
it.

One important thing to note is that server-wise there are now two parallel
gunicorn services that are stood up with them being cycled back and forth with
an nginx snippet. The reasoning is that we can severely reduce weird
side-effects during the time when code is being pushed up to a live server.