Cleans up local containers after role run
The role was previously allowing the helper containers created on the
Ansible controller, for preparing the build dependencies, to dangle
after the role completed. Added a final step to stop them and remove the
associated volumes.

In the process, I renamed pf_tracker_node -> django_stack_node for
consistency with the django_stack_venv container name.