Fix rsync capability for local code setup
This is hacky work-around to fix #28 - I continually hit issues using the built-in
synchronize task between two local directories and was unable to resolve. I
attempted to throw the develop version in place to solve this to no avail.
Normally I would file an upstream issue on this but I'm kind of short in time
and this task doesnt need to be idempotent soo I'm going to stick with the
`command` workaround for now.