Updates apt cache in venv builder container
The `apt` module logic that applies to the local container for
constructing the virtualenv did not have an update_cache parameter.
It's not possible to target the build container at the play-level, as a
form of override, since the build container is temporary, created and
destroyed within the lifecycle of a single role run.

Unfortunately it's not a simple matter to add config tests to validate
the new behavior: any packages listed in `django_stack_pkgs` will *not*
be present in the configured container after role execution—those
packages were installed only within the temporary build container.

Closes #30.