Makes Node builder Dockerfile configurable
Over in Secure The News, node-sass is used by gulp to build assets.
Node-sass must build bindings, so we need extra build dependencies.
Rather than add them here, in a generalized role used for multiple
projects, we can add it only to the STN provisioning config, as long as
the Dockerfile used for the Node logic can be overridden.