Forces fresh builds for local Node build container
Uses force=yes to rebuild a minimal image even if a matching tag was
already found locally. This is because the role now supports overriding
the Dockerfile used for building the NodeJS environment, so we don't
want a previous incarnation of the node builder to be in effect in a
custom environment. Forcing rebuilds every time slows down deployments
slightly, but at the cost of making them far more reliable.

Also sets keep_volumes=no on the Node container. In this specific
context, I believe the option is unwarranted: only local dirs are
mounted into the countainer. Still, we don't ever want previous volumes
sticking around: each role run should create a fresh volume and copy
everything in from scratch. Forcing volume removal ensures the build is
fresh every time.