Also updates venv dir name in set facts
The role configures facts dynamically in-line to handle switch-over
between alpha/beta instances. Unfortunately the role logic as written
does not support switching to new virtualenv paths entirely.

At first it looked as if forcing `django_first_install: True` in-line in
`tasks/setup.yml` would be sufficient, but that's not the case. Looks
like we'll need a force_rebuild_virtualenv var or similar—some way to
tell the role everything should be setup from scratch.