Restarts celery worker on config changes
The celery worker has its own config file, separate from the systemd
service file, and changes to that should trigger service bounces.
The logic is a bit convoluted here due to the a/b service naming, but in
short we want to halt the celery worker BEFORE handlers run to make sure
the correct services are referenced in all places.